Addressing some of the reviewer feedback.
