+++ geomesa-convert/geomesa-convert-common/src/main/scala/org/locationtech/geomesa/convert2/package.scala
@@ -11,8 +11,8 @@ package org.locationtech.geomesa
 import java.nio.charset.Charset
 
 import com.typesafe.config.Config
-import org.locationtech.geomesa.convert.ErrorMode.ErrorMode
-import org.locationtech.geomesa.convert.ParseMode.ParseMode
+import org.locationtech.geomesa.convert.Modes.ErrorMode
+import org.locationtech.geomesa.convert.Modes.ParseMode
 import org.locationtech.geomesa.convert.{EvaluationContext, SimpleFeatureValidator}
 import org.locationtech.geomesa.convert2.transforms.Expression
